+++ build.rs
@@ -27,7 +27,7 @@ fn main() {
         .filter(|line| !line.is_empty())
     {
         let (key, value) = line.split_at(line.find(" ").unwrap());
-        fmt::write(&mut output, format_args!("\t(\"{}\", r###\"{}\"###),\n", key, &value[1..])).unwrap();
+        fmt::write(&mut output, format_args!("\t(\"{}\", r#\"{}\"#),\n", key, &value[1..])).unwrap();
     }
 
     fmt::write(&mut output, format_args!("];\n")).unwrap();