+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All user visible changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/), as described
 for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/text/1105-api-evolution.md)
 
-## [0.3.0] - (pending)
+## [0.3.1] - (pending)
+
+ * Allow to specify named patterns when compiling, without inserting the definition beforehand.
+
+## [0.3.0] - 2017-09-13
 
  * `regex` has been switched to `onig` so we have full compatibility with all the other grok patterns.
  * Added `Grok::with_patterns()` which loads all the default patterns. `Grok::defalt()` also uses that now.